+++ programs/core/src/libraries/liquidity_math.rs
@@ -10,13 +10,13 @@ use anchor_lang::require;
 /// * `x` - The liquidity (L) before change
 /// * `y` - The delta (ΔL) by which liquidity should be changed
 ///
-pub fn add_delta(x: u64, y: i64) -> Result<u64, anchor_lang::error::Error> {
-    let z: u64;
+pub fn add_delta(x: u128, y: i128) -> Result<u128, anchor_lang::error::Error> {
+    let z: u128;
     if y < 0 {
-        z = x - (-y as u64);
+        z = x - (-y as u128);
         require!(z < x, ErrorCode::LiquiditySubValueErr);
     } else {
-        z = x + (y as u64);
+        z = x + (y as u128);
         require!(z >= x, ErrorCode::LiquidityAddValueErr);
     }
